+++ packages/classes/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.3.0] - 2020-07-12
 
 ### Added
+- `AvroPayload` to serialize and deserialize binary data.
 - A static `isPayload` method to the `Payload` class.
 - A static `isPayloadArray` method to the `PayloadArray` class.